+++ xarray/core/dataset.py
@@ -34,6 +34,8 @@ from .utils import (
     ensure_us_time_resolution, hashable, maybe_wrap_array, to_numeric)
 from .variable import IndexVariable, Variable, as_variable, broadcast_variables
 
+from ..coding.cftimeindex import _parse_array_of_cftime_strings
+
 # list of attributes of pd.DatetimeIndex that are ndarrays of time info
 _DATETIMEINDEX_COMPONENTS = ['year', 'month', 'day', 'hour', 'minute',
                              'second', 'microsecond', 'nanosecond', 'date',
@@ -1412,8 +1414,8 @@ class Dataset(Mapping, ImplementsDatasetReduce, DataWithCoords,
         """ Here we make sure
         + indexer has a valid keys
         + indexer is in a valid data type
-        * string indexers are cast to datetime64
-          if associated index is DatetimeIndex
+        + string indexers are cast to the appropriate date type if the 
+          associated index is a DatetimeIndex or CFTimeIndex
         """
         from .dataarray import DataArray
 
@@ -1435,10 +1437,12 @@ class Dataset(Mapping, ImplementsDatasetReduce, DataWithCoords,
             else:
                 v = np.asarray(v)
 
-                if ((v.dtype.kind == 'U' or v.dtype.kind == 'S')
-                    and isinstance(self.coords[k].to_index(),
-                                 pd.DatetimeIndex)):
-                    v = v.astype('datetime64[ns]')
+                if v.dtype.kind == 'U' or v.dtype.kind == 'S':
+                    index = self.indexes[k]
+                    if isinstance(index, pd.DatetimeIndex):
+                        v = v.astype('datetime64[ns]')
+                    elif isinstance(index, xr.CFTimeIndex):
+                        v = _parse_array_of_cftime_strings(v, index.date_type)
 
                 if v.ndim == 0:
                     v = as_variable(v)
@@ -3807,19 +3811,21 @@ class Dataset(Mapping, ImplementsDatasetReduce, DataWithCoords,
                              ' dimensional'.format(coord, coord_var.ndim))
 
         dim = coord_var.dims[0]
-        coord_data = coord_var.data
-        if coord_data.dtype.kind in 'mM':
-            if datetime_unit is None:
-                datetime_unit, _ = np.datetime_data(coord_data.dtype)
-            coord_data = to_numeric(coord_data, datetime_unit=datetime_unit)
+        if _contains_datetime_like_objects(coord_var):
+            if coord_var.dtype.kind in 'mM' and datetime_unit is None:
+                datetime_unit, _ = np.datetime_data(coord_var.dtype)
+            elif datetime_unit is None:
+                datetime_unit = 's'  # Default to seconds for cftime objects
+            coord_var = to_numeric(coord_var, datetime_unit=datetime_unit)
 
         variables = OrderedDict()
         for k, v in self.variables.items():
             if (k in self.data_vars and dim in v.dims and
                     k not in self.coords):
-                v = to_numeric(v, datetime_unit=datetime_unit)
+                if _contains_datetime_like_objects(v):
+                    v = to_numeric(v, datetime_unit=datetime_unit)
                 grad = duck_array_ops.gradient(
-                    v.data, coord_data, edge_order=edge_order,
+                    v.data, coord_var, edge_order=edge_order,
                     axis=v.get_axis_num(dim))
                 variables[k] = Variable(v.dims, grad)
             else: