+++ xarray/core/missing.py
@@ -9,6 +9,7 @@ import numpy as np
 import pandas as pd
 
 from . import rolling
+from .common import _contains_datetime_like_objects
 from .computation import apply_ufunc
 from .pycompat import iteritems
 from .utils import is_scalar, OrderedSet, to_numeric
@@ -407,13 +408,13 @@ def _floatize_x(x, new_x):
     x = list(x)
     new_x = list(new_x)
     for i in range(len(x)):
-        if x[i].dtype.kind in 'Mm':
+        if _contains_datetime_like_objects(x[i]):
             # Scipy casts coordinates to np.float64, which is not accurate
             # enough for datetime64 (uses 64bit integer).
             # We assume that the most of the bits are used to represent the
             # offset (min(x)) and the variation (x - min(x)) can be
             # represented by float.
-            xmin = np.min(x[i])
+            xmin = x[i].min()
             x[i] = to_numeric(x[i], offset=xmin, dtype=np.float64)
             new_x[i] = to_numeric(new_x[i], offset=xmin, dtype=np.float64)
     return x, new_x